% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{disco_color_scale}
\alias{disco_color_scale}
\title{Map disco scores to colors}
\usage{
disco_color_scale(x, lower = -100, upper = 100, int = 255, alpha = "66")
}
\arguments{
\item{x}{disco score}

\item{lower, upper}{lower and upper boundaries of the score}

\item{int}{how many colors in the palette}

\item{alpha}{transparency (string with two hexadecimal digits)}
}
\value{
character vector of colors corresponding to x
}
\description{
Map disco scores to colors
}
