% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_disco.R
\name{discoUI}
\alias{discoUI}
\alias{discoServer}
\title{Shiny Module – disco plots}
\usage{
discoUI(id, cntr_titles)

discoServer(
  id,
  cntr,
  annot = NULL,
  selcols = c("PrimaryID", "ENTREZ", "SYMBOL"),
  primary_id = "PrimaryID",
  gene_id = NULL
)
}
\arguments{
\item{id}{identifier of the shiny module (character vector)}

\item{cntr_titles}{character vector containing the IDs of the contrasts
(same as \code{names(cntr)}).}

\item{cntr}{named list of dataset contrast lists. Contrast data frames must
have the columns log2FoldChange and pvalue. For a single dataset, use
\code{list(default=cntr)}.}

\item{annot}{named list of dataset annotation data frames. Each annotation
data frame must have a column named "PrimaryID". For a single dataset,
use \code{list(default=annot)}.}

\item{selcols}{which column in the gene table when genes are selected
from the plot}

\item{primary_id}{name of the contrast data frame column with the primary IDs}

\item{gene_id}{must be a \code{reactiveValues} object. If not NULL, then
clicking on a gene identifier will modify this object (possibly
triggering an event in another module).}
}
\value{
Returns a reactive expression returning the ID of the activated gene
}
\description{
Shiny Module – disco plots
}
\examples{
if(interactive()) {
    cntr1 <- data.frame(log2FoldChange=rnorm(5000),
                        pvalue=runif(5000),
                        PrimaryID=paste0("ID", 1:5000))
    cntr2 <- data.frame(log2FoldChange=cntr1$log2FoldChange + 
                                       rnorm(5000),
                        pvalue=runif(5000) * cntr1$pvalue,
                        PrimaryID=paste0("ID", 1:5000))
    cntr1$SecondaryID <- sample(cntr1$PrimaryID, replace=TRUE)
    cntr2$SecondaryID <- sample(cntr2$PrimaryID, replace=TRUE)
    rownames(cntr2) <- paste0("ID", 1:5000)
    cntr <- list("Contrast 1"=cntr1, "Contrast 2"=cntr2)
    annot <- data.frame(PrimaryID=paste0("ID", 1:5000), stringsAsFactors=FALSE)
    shinyApp(ui=fluidPage(discoUI("disco", names(cntr))),
             server=function(input, output, session) {
                discoServer("disco", list(default=cntr), annot=list(default=annot))
             })
}
}
