% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{figsizeInput}
\alias{figsizeInput}
\title{Figure size input}
\usage{
figsizeInput(
  id,
  label = "Figure size",
  choices = c("800x800", "600x600", "600x800", "800x600", "1200x600", "1200x800",
    "600x1200", "800x1200", "1200x1200"),
  selected = NULL,
  custom = TRUE,
  ...
)
}
\arguments{
\item{id}{ID of the element}

\item{label}{Label for the input}

\item{choices}{Which figure size choices (in pixels)}

\item{selected}{Default figure size}

\item{custom}{Whether custom fig sizes are allowed}

\item{...}{further arguments passed to selectizeInput}
}
\description{
Figure size input
}
