% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_browser.R
\name{gene_browser}
\alias{gene_browser}
\title{Launch a browser of DE analysis results}
\usage{
gene_browser(
  x,
  annot_linkout = NULL,
  cols_to_hide = c("stat", "lfcSE", "symbol", "entrez"),
  primary_id = "PrimaryID"
)
}
\arguments{
\item{x}{object holding the DE analysis results. List containing the
elements \code{contrasts}, \code{covariates}, \code{expression} and \code{annotation}.}

\item{annot_linkout}{a list; see Details.}

\item{cols_to_hide}{columns in the contrasts data frames to hide in the table.
Default is for DESeq2 derived contrasts.}

\item{primary_id}{The name of the column in \code{contrasts} and \code{annotation}
which holds the primary, unique identifier of genes.}
}
\value{
does not return a value
}
\description{
Launch a simple shiny-based browser of DE analysis results
}
\details{
Creates a shiny app, which allows to show gene expression
profiles.
}
\examples{
if(interactive()) {
  data(C19)
  annot_linkout <- list(
    SYMBOL="https://www.genecards.org/cgi-bin/carddisp.pl?gene=\%s",
    ENTREZ="https://www.ncbi.nlm.nih.gov/gene/?term=\%s"
  )
  gene_browser(C19, annot_linkout=annot_linkout)
}
}
