% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_gene_browser_table.R
\name{geneBrowserTableUI}
\alias{geneBrowserTableUI}
\alias{geneBrowserTableServer}
\title{Shiny Module – gene browser table selection}
\usage{
geneBrowserTableUI(id, cntr_titles)

geneBrowserTableServer(
  id,
  cntr,
  annot,
  annot_linkout = NULL,
  primary_id = "PrimaryID",
  gene_id = NULL,
  cols_to_hide = c("stat", "lfcSE", "symbol", "entrez")
)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{named character vector for contrast choices}

\item{cntr}{named list of dataset contrast lists}

\item{annot}{named list of dataset annotation data frames containing column
'PrimaryID' (or another specified by the parameter \code{primary_id})}

\item{annot_linkout}{a list; see Details.}

\item{primary_id}{name of the column which holds the primary identifiers}

\item{gene_id}{must be either NULL or a \code{reactiveValues} object. If not NULL, then
a button is displayed; clicking on it will modify the \code{gene_id} reactive
value (possibly triggering an event in another module).}

\item{cols_to_hide}{columns in the contrasts data frames to hide in the table.
Default is for DESeq2 derived contrasts.}
}
\value{
\code{geneBrowserTableServer} returns NULL. \code{geneBrowserTableUI}
returns the interface.
}
\description{
Shiny Module – gene browser table selection
}
\details{
The basic data structure is a named list of datasets. Each dataset is a named
list of contrast data frames. For a single dataset, use
\verb{list(default=<contrast list>)}. The contrast data frames \emph{must} contain a
column called "PrimaryID" (this identifier can be changed with the parameter
\code{primary_id}). This is necessary in order to link the table rows with e.g.
plotting genes with \code{geneBrowserPlotServer}.

Log2 fold changes must be stored in a column called
"log2FoldChange", and p-values in a column called "padj". These are the
default column names returned by DESeq2.

The linkout feature (parameter \code{annot_linkout}) allows to define how the
different columns from the annotation data frame are represented as
linkouts to external data bases.

The parameter \code{annot_linkout} is a named list. Names must correspond to
columns from the annotation data frame. The elements of the list are
character strings containing URLs with the \verb{\%s} placeholder. For
example, if the column \code{ENSEMBL} contains ENSEMBL identifiers, you can
link out by specifying

\if{html}{\out{<div class="sourceCode">}}\preformatted{annot_linkout=list(ENSEMBL="https://www.ensembl.org/id/\%s")
}\if{html}{\out{</div>}}

The parameter \code{gene_id} is used to communicate with other components of
the shiny app, e.g. with another module (such as geneBrowserPlotServer).
If not NULL, it must be a reactiveValues object.
When the gene_id is not NULL, colored buttons appear next to the gene ID
in the table. Clicking on them causes changing the information in
\code{gene_id$id} (ID of the gene) and \code{gene_id$ds} (dataset or "default" if
there is only a single data set).
}
\examples{
data(C19)

if(interactive()) {
  ui <- fluidPage(
           geneBrowserTableUI("gb", names(C19$contrasts))
        )

  server <- function(input, output) {
    geneBrowserTableServer("gb", cntr=list(default=C19$contrasts),
                                 annot=list(default=C19$annotation))
  }

  shinyApp(ui, server)
}

## Using the gene_id reactive value
## Clicking on the button displays the gene ID
if(interactive()) {
  ui <- fluidPage(
           geneBrowserTableUI("gb", names(C19$contrasts)),
           textOutput("gene")
        )

  server <- function(input, output) {
    gene_id <- reactiveValues()
    geneBrowserTableServer("gb", gene_id=gene_id,
                                 cntr=list(default=C19$contrasts),
                                 annot=list(default=C19$annotation))
    output$gene <- 
      renderText(sprintf("gene ID: \%s, data set: \%s", gene_id$id, gene_id$ds))
  }

  shinyApp(ui, server)
}
}
