% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tmod_panelplot.R
\name{gg_panelplot_chunk}
\alias{gg_panelplot_chunk}
\title{Generate code for a tmod panel plot}
\usage{
gg_panelplot_chunk(
  res,
  pie,
  auc_thr = 0.5,
  q_thr = 0.05,
  filter_row_q = 0.01,
  filter_row_auc = 0.65,
  q_cutoff = 1e-12,
  label_angle = 45,
  cleanup = TRUE,
  add_ids = TRUE,
  env_map = NULL
)
}
\arguments{
\item{res}{a list with tmod results (each element of the list is a data
frame returned by a tmod test function)}

\item{pie}{a list with summaries of significantly DE genes by gene set.
Each a element of the list is a matrix returned by tmodDecideTests.}

\item{auc_thr}{gene sets enrichments with AUC below \code{auc_thr} will not be shown}

\item{q_thr}{gene sets enrichments with q (adjusted P value) above \code{q_thr} will not be shown}

\item{filter_row_q}{Gene sets will only be shown if at least in one
contrast q is smaller than \code{filter_row_q}}

\item{filter_row_auc}{Gene sets will only be shown if at least in one
contrast AUC is larger than \code{filter_row_q}}

\item{q_cutoff}{Any q value below \code{q_cutoff} will be considered equal to
\code{q_cutoff}}

\item{label_angle}{The angle at which column labels are shown}

\item{cleanup}{attempt to clean up gene set titls}

\item{add_ids}{add IDs of gene sets to their titles on the plot}

\item{env_map}{optional list mapping \code{res} and \code{pie} to names used in the generated code.}
}
\value{
a list with metadata and generated code
}
\description{
Generate readable R code that mirrors \code{\link[=gg_panelplot]{gg_panelplot()}} without calling it.
}
