% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_disco}
\alias{plot_disco}
\title{Create a disco plot}
\usage{
plot_disco(
  contrast1,
  contrast2,
  lower = -100,
  upper = 100,
  show_top_labels = 0,
  top_labels_both = TRUE,
  annot = NULL,
  alpha = 0.5,
  disco = NULL,
  by = 0,
  primary_id = "PrimaryID",
  label_col = "SYMBOL"
)
}
\arguments{
\item{contrast1, contrast2}{data frames with rownames corresponding to
IDs (they don't need to be in the same order) and columns \code{log2FoldChange}
and \code{pvalue}.}

\item{lower, upper}{lower and upper boundaries for coloring of the score}

\item{show_top_labels}{sort the genes by descending absolute disco score and show top N labels}

\item{top_labels_both}{should top labels from both negative and positive
disco scores be shown, or only for the absolute top, whether only negative
or only positive or both?}

\item{annot}{annotation object returned by \code{get_annot()} or any other data
frame with columns "PrimaryID" and "SYMBOL"}

\item{alpha}{transparency}

\item{disco}{result of \code{disco_score}; if provided, it will be used to
avoid unnecessary computations}

\item{by}{column by which the contrast data frames should be merged
(passed to \code{merge}). Default: merge by row names}

\item{primary_id}{the name which should be assigned to the identifier
column which results from the merge}

\item{label_col}{name of the column in \code{annot} which should be used for
labels}
}
\value{
a ggplot object (plot)
}
\description{
Make a concordance / discordance plots for two contrasts
}
\details{
A concordance / discordance plot is simply a log~2~ FC vs log~2~ FC
comparison between two contrasts.
Disco score is a heuristic score
reflecting the strength of similarity or dissimilarity between two
log~2~ FC values weighted by the corresponding p-values and given by the
formula log2FC.x * log2FC.y * (-log10(pval.x) - log10(pval.y)).

To manually select which labels to show, use a truncated \code{annot} object
– remove all rows you don't want to show, and set the show_top_labels
parameter to \code{Inf}.
}
\examples{
## Generate example data
c1 <- data.frame(log2FoldChange=rnorm(5000, sd=2))
c1$pvalue <- pnorm(abs(c1$log2FoldChange), sd=2, lower.tail=FALSE)
c2 <- data.frame(log2FoldChange=c1$log2FoldChange + rnorm(5000, sd=3))
c2$pvalue <- pnorm(abs(c2$log2FoldChange), sd=2, lower.tail=FALSE)

## Example disco plot
plot_disco(c1, c2)
}
