% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_disco_chunk}
\alias{plot_disco_chunk}
\title{Create code for a disco plot}
\usage{
plot_disco_chunk(
  contrast1,
  contrast2,
  lower = -100,
  upper = 100,
  show_top_labels = 0,
  top_labels_both = TRUE,
  annot1 = NULL,
  annot2 = NULL,
  alpha = 0.5,
  disco = NULL,
  by = 0,
  primary_id = "PrimaryID",
  label_col = "SYMBOL",
  label_sel = NULL,
  env_map = NULL
)
}
\arguments{
\item{contrast1, contrast2}{data frames with rownames corresponding to
IDs (they don't need to be in the same order) and columns \code{log2FoldChange}
and \code{pvalue}.}

\item{lower, upper}{lower and upper boundaries for coloring of the score}

\item{show_top_labels}{sort the genes by descending absolute disco score and show top N labels}

\item{top_labels_both}{should top labels from both negative and positive
disco scores be shown, or only for the absolute top, whether only negative
or only positive or both?}

\item{annot1, annot2}{annotation data frames for the two contrasts}

\item{alpha}{transparency}

\item{disco}{result of \code{disco_score}; if provided, it will be used to
avoid unnecessary computations}

\item{by}{column by which the contrast data frames should be merged
(passed to \code{merge}). Default: merge by row names}

\item{primary_id}{the name which should be assigned to the identifier
column which results from the merge}

\item{label_col}{name of the column in \code{annot} which should be used for
labels}

\item{label_sel}{identifiers of labels to be shown on the plot}

\item{env_map}{optional list mapping \code{contrast1}, \code{contrast2}, \code{annot1},
\code{annot2}, and \code{disco} to names used in the generated code.}
}
\value{
a list with metadata and generated code
}
\description{
Create readable R code that mirrors \code{\link[=plot_disco]{plot_disco()}} without calling it.
}
