% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ly_pca}
\alias{plot_ly_pca}
\title{Create a PCA plot with plotly}
\usage{
plot_ly_pca(mtx, covariate_data, threeD = TRUE, cov_default = NULL)
}
\arguments{
\item{mtx}{PCA matrix (typically \code{prcomp(...)$x})}

\item{covariate_data}{data frame with covariates to display (character, factor or continuous)}

\item{threeD}{if FALSE, a 2D plot will be produced}

\item{cov_default}{default covariate names (must be names of columns of \code{covariate_data})}
}
\value{
the plotly object
}
\description{
Generate a PCA 3D or 2D plot with plotly
}
