% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_pca.R
\name{plot_pca_shiny}
\alias{plot_pca_shiny}
\title{Simple PCA plot in a browser}
\usage{
plot_pca_shiny(x, covar, colorBy = NULL, symbolBy = NULL)
}
\arguments{
\item{x}{matrix or data frame for PCA}

\item{covar}{covariate file}

\item{colorBy}{name of the covariate column for coloring}

\item{symbolBy}{name of the covariate column for chosing symbols}
}
\description{
Simple PCA plot in a browser using shiny
}
\examples{
if(interactive()) {
   plot_pca_shiny(iris[,1:4], iris[,5,drop=FALSE])
}
}
