% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_disco.R
\name{discoUI}
\alias{discoUI}
\alias{discoServer}
\title{Shiny Module – disco plots}
\usage{
discoUI(id, cntr_titles)

discoServer(
  id,
  cntr,
  annot = NULL,
  selcols = c("PrimaryID", "ENTREZ", "SYMBOL"),
  primary_id = "PrimaryID",
  gene_id = NULL
)
}
\arguments{
\item{id}{identifier of the shiny module (character vector)}

\item{cntr_titles}{character vector containing the IDs of the contrasts
(same as \code{names(cntr)}).}

\item{cntr}{list of data frames containing the contrast information.
Data frames must have the columns log2FoldChange and pvalue. Rownames of
the data frames should be IDs.}

\item{annot}{Annotation data frame. The annotation data frame must have
a column named "PrimaryID" which corresponds to the rownames of the data
frames in the \code{cntr} list.}

\item{selcols}{which column in the gene table when genes are selected
from the plot}

\item{primary_id}{name of the contrast data frame column with the primary IDs}

\item{gene_id}{must be a \code{reactiveValues} object. If not NULL, then
clicking on a gene identifier will modify this object (possibly
triggering an event in another module).}
}
\value{
Returns a reactive expression returning the ID of the activated gene
}
\description{
Shiny Module – disco plots
}
\examples{
if(interactive()) {
   cntr1 <- data.frame(log2FoldChange=rnorm(5000),
                       pvalue=runif(5000))
   rownames(cntr1) <- paste0("ID", 1:5000)
   cntr2 <- data.frame(log2FoldChange=cntr1$log2FoldChange + 
                                      rnorm(5000),
                       pvalue=runif(5000) * cntr1$pvalue)
   rownames(cntr2) <- paste0("ID", 1:5000)
   cntr <- list("Contrast 1"=cntr1, "Contrast 2"=cntr2)
   shinyApp(ui=fluidPage(discoUI("disco", names(cntr))),
            server=function(input, output, session) {
               discoServer("disco", cntr)
            })
}
}
