% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{geneBrowserTableUI}
\alias{geneBrowserTableUI}
\alias{geneBrowserTableServer}
\title{Shiny Module – gene browser table selection}
\usage{
geneBrowserTableUI(id, cntr_titles)

geneBrowserTableServer(
  id,
  cntr,
  annot,
  annot_linkout = NULL,
  primary_id = "PrimaryID",
  gene_id = NULL
)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{named character vector for contrast choices}

\item{cntr}{a list of data frames containing the DE analysis results, or
a list of lists of data frames}

\item{annot}{annotation data frame containing column 'PrimaryID' (
or another specified by the parameter \code{primary_id})
corresponding to the rownames of the contrast data frames}

\item{annot_linkout}{a list; see Details.}

\item{primary_id}{name of the column which holds the primary identifiers}

\item{gene_id}{must be a \code{reactiveValues} object. If not NULL, then
clicking on a gene identifier will modify this object (possibly
triggering an event in another module).}
}
\value{
reactive value containing the gene ID
}
\description{
Shiny Module – gene browser table selection
}
\details{
The basic data set structure that this module takes is a named list of data
frames. These data frames will be shown in the browser when the specific
contrast (corresponding to a name in the list) is selected from the
configuration sidebar. The data frames \emph{must} contain a column called
"PrimaryID" (this identifier can be changed with the parameter
\code{primary_id}). This is necessary in order to link the table rows with
e.g. plotting genes with \code{geneBrowserPlotServer}.

Log2 fold changes must be stored in a column called
"log2FoldChange", and p-values in a column called "padj". These are the
default column names returned by DESeq2.

Alternatively, tmodBrowserTableServer takes a list of lists of data
frames; that is, it allows to group the results of differential gene
analysis.

The linkout feature (parameter \code{annot_linkout}) allows to define how the
different columns from the annotation data frame are represented as
linkouts to external data bases.

The parameter \code{annot_linkout} is a named list. Names must correspond to
columns from the annotation data frame. The elements of the list are
character strings containing URLs with the \verb{\%s} placeholder. For
example, if the column \code{ENSEMBL} contains ENSEMBL identifiers, you can
link out by specifying\preformatted{annot_linkout=list(ENSEMBL="https://www.ensembl.org/id/\%s")
}
}
