% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tmod_browser.R
\name{tmodBrowserTableUI}
\alias{tmodBrowserTableUI}
\alias{tmodBrowserTableServer}
\title{Shiny Module – tmod results browser table selection}
\usage{
tmodBrowserTableUI(id, cntr_titles, upset_pane = FALSE)

tmodBrowserTableServer(
  id,
  tmod_res,
  gs_id = NULL,
  multilevel = FALSE,
  tmod_dbs = NULL
)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{possibly named character vector with contrast names}

\item{upset_pane}{if TRUE, UI for the upset plot will be created}

\item{tmod_res}{results of tmod analysis, returned by \code{get_tmod_res}}

\item{gs_id}{a list of reactive values (returned by \code{reactiveValues()}), including
dataset (\code{ds}), gene set ID (\code{id}), contrast id (\code{cntr}), database ID
(\code{db}) and sorting mode (\code{sort}). If \code{mod_id} is not \code{NULL}, these
reactive values will be populated, possibly triggering an action in
another shiny module.}

\item{multilevel}{if TRUE, the results are grouped in data sets}

\item{tmod_dbs}{list of lists of tmod database objects (or list of lists
of list of tmod database object in multilevel mode). If NULL, upset
plots cannot be generated.}
}
\description{
Shiny Module – tmod results browser table selection
}
