% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tmod_panelplot.R
\name{tmodPanelPlotUI}
\alias{tmodPanelPlotUI}
\alias{tmodPanelPlotServer}
\title{Shiny module displaying tmod panel plots}
\usage{
tmodPanelPlotUI(id, datasets = NULL)

tmodPanelPlotServer(
  id,
  cntr,
  tmod_res,
  tmod_dbs,
  tmod_map,
  gs_id = NULL,
  annot = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{datasets}{if there are multiple data sets, this character vector
defines what they are to show an approppriate selector in the UI}

\item{cntr}{list of data frames with the results of differential
expression analysis. Rownames must correspond to the 'PrimaryID' column
of data frame annot.}

\item{tmod_res}{list of tmod gene set enrichment analysis
results. See Details.}

\item{tmod_dbs}{named list of tmod databases, see Details.}

\item{tmod_map}{mapping between the PrimaryIDs from the contrasts and
the gene IDs from the gene set databases.}

\item{gs_id}{a "reactive values" object (returned by \code{reactiveValues()}), including
dataset (\code{ds}), gene set ID (\code{id}), contrast id (\code{cntr}), database ID
(\code{db}) and sorting mode (\code{sort}). If \code{mod_id} is not \code{NULL}, these
reactive values will be populated, possibly triggering an action in
another shiny module.}

\item{annot}{data frame containing at least the column 'PrimaryID'}
}
\value{
Returns a reactive value which is a list with elements
\code{contrast} and \code{id}.
}
\description{
Shiny module displaying tmod panel plots
}
\details{
Tmod results, mapping and databases.

The \code{tmod_res} object is a nested list with the following levels:
\itemize{
\item top level are the contrasts. \code{names(tmod_res)} must be (set)equal to
\code{names(cntr)}.
\item next level are the names of tmod databases. \code{names(tmod_res[[1]])}
must be (set)equal to \code{names(tmod_dbs)}
}
}
