% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_volcano.R
\name{volcanoUI}
\alias{volcanoUI}
\alias{volcanoServer}
\title{Shiny module for displaying volcano plots}
\usage{
volcanoUI(id, datasets = NULL, lfc_thr = 1, pval_thr = 0.05)

volcanoServer(
  id,
  cntr,
  lfc_col = "log2FoldChange",
  pval_col = "padj",
  primary_id = "PrimaryID",
  annot = NULL,
  gene_id = NULL,
  selection = NULL,
  ui_config = NULL,
  annot_show = c("SYMBOL", "ENTREZID")
)
}
\arguments{
\item{id}{module identifier (same as the one passed to volcanoUI)}

\item{datasets}{character vector specifying datasets}

\item{lfc_thr}{default lfc threshold}

\item{pval_thr}{default p-value threshold}

\item{cntr}{either a named list of data frames, each being the results
of differential expression analysis for one contrast, or a list of data
sets, each data set being a named list of data frames.}

\item{lfc_col, pval_col}{names of the columns in the contrast data
frames which hold the log2 fold changes and p-values, respectively}

\item{primary_id}{name of the primary ID column in contrasts and
annotation data frame.}

\item{annot}{data frame with gene annotations (containing at least the
column specified with the \code{primary_id} parameter) or (if there are
multiple data sets) a named list of such data frames. Names of this list
must match the names of the \code{cntr} list.}

\item{gene_id}{must be a \code{reactiveValues} object. If not NULL, then
clicking on a gene identifier will modify this object (possibly
triggering an event in another module).}

\item{selection}{Optional \code{reactiveValues()} object with an \code{ids}
element populated with the primary IDs from the current plot selection when
the user clicks the \code{Show} button. If \code{NULL}, the button is not
shown and no external state is updated.}

\item{ui_config}{Optional list configuring UI text. Supported keys:
\code{show_button_label} for the label shown on the \code{Show} button.}

\item{annot_show}{which columns from the annotation data frame should be
shown when mouse hovers over a gene}
}
\description{
Shiny module for displaying volcano plots
}
\examples{
annot <- data.frame(
  PrimaryID=paste0("g", 1:6),
  SYMBOL=LETTERS[1:6],
  stringsAsFactors=FALSE
)
exprs <- matrix(
  rnorm(24),
  nrow=6,
  dimnames=list(annot$PrimaryID, paste0("s", 1:4))
)
cntr <- list(
  contrast_a=data.frame(
    PrimaryID=annot$PrimaryID,
    log2FoldChange=rnorm(6),
    padj=runif(6),
    stringsAsFactors=FALSE
  )
)
covar <- data.frame(
  SampleID=colnames(exprs),
  group=c("A", "A", "B", "B"),
  stringsAsFactors=FALSE
)

if(interactive()) {
  ui <- fluidPage(
    fluidRow(
      column(width=6, volcanoUI("vol")),
      column(width=6, heatmapUI("hm"))
    )
  )
  server <- function(input, output, session) {
    selection <- reactiveValues(ids=character(0))
    volcanoServer("vol", cntr=cntr, annot=annot, selection=selection)
    heatmapServer(
      "hm",
      annot=annot,
      exprs=exprs,
      cntr=cntr,
      covar=covar,
      selection=selection
    )
  }
  shinyApp(ui, server)
}

}
