% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_volcano.R
\name{volcanoUI}
\alias{volcanoUI}
\alias{volcanoServer}
\title{Shiny module for displaying volcano plots}
\usage{
volcanoUI(id, datasets = NULL, lfc_thr = 1, pval_thr = 0.05)

volcanoServer(
  id,
  cntr,
  lfc_col = "log2FoldChange",
  pval_col = "padj",
  primary_id = "PrimaryID",
  annot = NULL,
  gene_id = NULL
)
}
\arguments{
\item{id}{module identifier (same as the one passed to volcanoUI)}

\item{datasets}{character vector specifying datasets}

\item{lfc_thr}{default lfc threshold}

\item{pval_thr}{default p-value threshold}

\item{cntr}{either a named list of data frames, each being the results
of differential expression analysis for one contrast, or a list of data
sets, each data set being a named list of data frames.}

\item{lfc_col, pval_col}{names of the columns in the contrast data
frames which hold the log2 fold changes and p-values, respectively}

\item{primary_id}{name of the primary ID column in contrasts and
annotation data frame}

\item{annot}{data frame with gene annotations (containing at least the
column specified with the \code{primary_id} parameter) or (if there are
multiple data sets) a named list of such data frames. Names of this list
must match the names of the \code{cntr} list.}

\item{gene_id}{must be a \code{reactiveValues} object. If not NULL, then
clicking on a gene identifier will modify this object (possibly
triggering an event in another module).}
}
\description{
Shiny module for displaying volcano plots
}
